refactor: update calls to area_to_loadzone function
